% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_int.R
\name{age}
\alias{age}
\alias{age_int}
\alias{age_months}
\title{Calculate the age in years between 2 dates}
\usage{
age_int(from, to)

age_months(from, to)
}
\arguments{
\item{from}{Date. Birthday.}

\item{to}{Date. Day when age is computed.}
}
\value{
The number of full years between the 2 dates.
}
\description{
Calculate the age in years between 2 dates
}
