% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_diagnosis.R
\name{categorize_diagnosis}
\alias{categorize_diagnosis}
\title{Extract diagnosis for recoding}
\usage{
categorize_diagnosis(
  data,
  stcs,
  .patientkey = "patientkey",
  .startdate = NULL,
  .stopdate = NULL,
  .is_pre = NULL
)
}
\arguments{
\item{data}{data.frame. A data frame containing the variables describe in \code{.patientkey} and \code{.date}}

\item{stcs}{list. The stcs object.}

\item{.patientkey}{chr. The column name containing patientkey.}

\item{.startdate}{chr. The column name containing the starting date.}

\item{.stopdate}{chr. The column name containing the stop date.}

\item{.is_pre}{lgl. Level of the is_pre variable filter. Default is \code{.is_pre = c(TRUE,FALSE)}.}
}
\value{
a data frame containing \code{"disease_category"}, \code{"is_pre"}, \code{"patdiagnosis"}, \code{"n_occurence"}, \code{"range"}.
}
\description{
Extract diagnosis for recoding
}
