% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_medication.R
\name{categorize_medication}
\alias{categorize_medication}
\title{Extract drug data for recoding}
\usage{
categorize_medication(
  data,
  stcs,
  .drug_category = c("drug_idpro", "drug_ind", "drug_is", "drug_other"),
  .patientkey = "patientkey",
  .date = "date",
  .days_range = c(-Inf, Inf)
)
}
\arguments{
\item{data}{data.frame. A data frame containing the variables describe in \code{.patientkey} and \code{.date}}

\item{stcs}{list. The stcs object.}

\item{.drug_category}{chr. The drug category to filter. Several choice between \code{"drug_idpro"}, \code{"drug_ind"}, \code{"drug_is"}, \code{"drug_other"}}

\item{.patientkey}{chr. The column name containing patientkey}

\item{.date}{chr. The column name containign the date}

\item{.days_range}{numeric. A vector of length selecting the date range around \code{date} to look for medication.}
}
\value{
a data frame containing \code{"drug_category"}, \code{"name"}, \code{"comment"}, \code{"code"}, \code{"n_occurence"}, \code{"range"} describing the selected drugs and their occurrence.
}
\description{
Extract drug data for recoding
}
