% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_organevent.R
\name{categorize_organevent}
\alias{categorize_organevent}
\title{Extract organ event data for recoding}
\usage{
categorize_organevent(
  data,
  stcs,
  .organevent_category = c("Allograftdisease", "Complications", "Underlyingdisease"),
  .startdate = NULL,
  .stopdate = NULL,
  .organkey = "organkey"
)
}
\arguments{
\item{data}{data.frame. A data frame containing the variables describe in \code{.patientkey} and \code{.date}}

\item{stcs}{list. The stcs object.}

\item{.organevent_category}{chr. The event category to filter. Several choice between \code{"Allograftdisease"}, \code{"Complications"}, \code{"Underlyingdisease"}}

\item{.startdate}{chr. The column name containing the starting date.}

\item{.stopdate}{chr. The column name containing the stop date.}

\item{.organkey}{chr. The column name containing organkey}
}
\value{
a data frame containing \code{"organevent_category"}, \code{"organevent"}, \code{"organevent_comment"}, \code{"n_occurence"}, \code{"range"}.
}
\description{
Extract organ event data for recoding
}
