% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_organintervention.R
\name{categorize_organintervention}
\alias{categorize_organintervention}
\title{Extract interventions data for recoding}
\usage{
categorize_organintervention(
  data,
  stcs,
  .startdate = NULL,
  .stopdate = NULL,
  .organkey = "organkey"
)
}
\arguments{
\item{data}{data.frame. A data frame containing the variables describe in \code{.patientkey} and \code{.date}}

\item{stcs}{list. The stcs object.}

\item{.startdate}{chr. The column name containing the starting date.}

\item{.stopdate}{chr. The column name containing the stop date.}

\item{.organkey}{chr. The column name containing organkey}
}
\value{
a data frame containing \code{"txint"}, \code{"txint_comment"},, \code{"organevent"} \code{"organevent_comment"}, \code{"n_occurence"}, \code{"range"}.
}
\description{
Extract interventions data for recoding
}
\details{
the date of the organ-event is use when the intervention date is NA.
}
