% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_otherdisease.R
\name{categorize_otherdisease}
\alias{categorize_otherdisease}
\title{Extract Other disease for recoding}
\usage{
categorize_otherdisease(
  data,
  stcs,
  .patientkey = "patientkey",
  .date = "date",
  .is_pre = NULL,
  .days_range = c(-Inf, Inf)
)
}
\arguments{
\item{data}{data.frame. A data frame containing the variables describe in \code{.patientkey} and \code{.date}}

\item{stcs}{list. The stcs object.}

\item{.patientkey}{chr. The column name containing patientkey}

\item{.date}{chr. The column name containign the date}

\item{.is_pre}{lgl. Level of the is_pre variable filter. Default is \code{.is_pre = c(TRUE,FALSE)}.}

\item{.days_range}{numeric. A vector of length selecting the date range around \code{date} to look for medication.}
}
\value{
a data frame containing \code{"disease_category"}, \code{"is_pre"}, \code{"patdiagnosis"}, \code{"n_occurence"}, \code{"range"}.
}
\description{
Extract Other disease for recoding
}
