% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_pathogenspecies.R
\name{categorize_pathogenspecies}
\alias{categorize_pathogenspecies}
\title{Extract infection pathogen species data for recoding}
\usage{
categorize_pathogenspecies(
  data,
  stcs,
  .diseasekey = "diseasekey",
  add_infsite = c("no", "all", "other", "other+na")
)
}
\arguments{
\item{data}{data.frame. A data frame containing the variables describe in \code{.patientkey} and \code{.date}}

\item{stcs}{list. The stcs object.}

\item{.diseasekey}{chr. The column name containing diseasekey}

\item{add_infsite}{lgl. Should the infection site be included for recoding. Default is \code{"no"}. Either \code{c("no","all","other","other+na")}.}
}
\value{
a data frame containing \code{"patdiagnosis"}, \code{"pathogen_species"}, \code{"comment"}, \code{"n_occurence"}.
}
\description{
Extract infection pathogen species data for recoding
}
\details{
when \code{add_infsite = "no"}, the infection site are kept NA's.
when \code{add_infsite = "yes"}, the infection site is added for all pathogens.
when \code{add_infsite = "other"}, the infection site is added only when \code{"pathogen_species"} is "Other" (eg: one of \code{c("Other bacteria","Other fungi","Other parasites","Other viruses")}).
when \code{add_infsite = "other+na"}, the infection site is added only when \code{"pathogen_species"} is "Other" and \code{"comment"} is NA.
}
