% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_rejectiontreatment.R
\name{categorize_rejectiontreatment}
\alias{categorize_rejectiontreatment}
\title{Extract rejection treatment data for recoding}
\usage{
categorize_rejectiontreatment(data, stcs, .biorejkey = "biorejkey")
}
\arguments{
\item{data}{data.frame. A data frame containing the variables describe in \code{.patientkey} and \code{.date}}

\item{stcs}{list. The stcs object.}

\item{.biorejkey}{chr. The column name containing biorejkey}
}
\value{
a data frame containing \code{"drug_category"}, \code{"name"}, \code{"comment"}, \code{"code"}, \code{"n_occurence"}, \code{"range"}.
}
\description{
Extract rejection treatment data for recoding
}
