% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_treatment.R
\name{categorize_treatment}
\alias{categorize_treatment}
\title{Extract treatment data for recoding}
\usage{
categorize_treatment(
  data,
  stcs,
  .drug_category = c("drug_idpro", "drug_ind", "drug_is", "drug_other"),
  .patientkey = "patientkey",
  .startdate = NULL,
  .stopdate = NULL
)
}
\arguments{
\item{data}{data.frame. A data frame containing the variables describe in \code{.patientkey} and \code{.date}}

\item{stcs}{list. The stcs object.}

\item{.drug_category}{chr. The drug category to filter. Several choice between \code{"drug_idpro"}, \code{"drug_ind"}, \code{"drug_is"}, \code{"drug_other"}}

\item{.patientkey}{chr. The column name containing patientkey}

\item{.startdate}{chr. The column name containing the starting date.}

\item{.stopdate}{chr. The column name containing the stop date.}
}
\value{
a data frame containing \code{"drug_category"}, \code{"name"}, \code{"comment"}, \code{"code"}, \code{"n_occurence"}, \code{"range"}.
}
\description{
Extract treatment data for recoding
}
