% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_value.R
\name{closest_value}
\alias{closest_value}
\title{Closest values}
\usage{
closest_value(x, target = round(c(0, 186, 365.25 * c(1:25))))
}
\arguments{
\item{x}{num. Input to derived closest values.}

\item{target}{num. Target values.}
}
\description{
Compute the closest vlues from target values.
}
\examples{

closest_value(1:10, c(0, 5, 10))

}
