% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_key.R
\name{data_key}
\alias{data_key}
\alias{data_patientkey}
\alias{data_organkey}
\alias{data_soaskey}
\title{Get STCS key}
\usage{
data_patientkey(stcs)

data_organkey(stcs)

data_soaskey(stcs)
}
\arguments{
\item{stcs}{list. A list of STCS tables}
}
\value{
A data.frame with the requested key.
}
\description{
Get STCS key
}
