% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detect}
\alias{detect}
\alias{detect_from}
\alias{detect_by}
\title{Detect if the tables/key can be automatically retrieve}
\usage{
detect_from(stcs, .var)

detect_by(stcs, .var, from = NULL)
}
\arguments{
\item{stcs}{A list containing the STCS data frame.}

\item{.var}{chr. The name of variables to add.}

\item{from}{chr. The name of the dataset (optional)}
}
\value{
A data frame with the new column.
}
\description{
Detect if the tables/key can be automatically retrieve
}
