% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{egfr_2021}
\alias{egfr_2021}
\title{eGFR}
\usage{
egfr_2021(crea_mumoll, age, sex)
}
\arguments{
\item{crea_mumoll}{Numeric. Creatinin in \eqn{\mu}mol/l}

\item{age}{Numeric. Age in years}

\item{sex}{Character. Sex either "Male" or "Female"}
}
\value{
eGFR in mL/min/1.73m\eqn{^2} as described in https://www.kidney.org/content/ckd-epi-creatinine-equation-2021. The conversion factor of creatinin is 1 (\eqn{\mu}mol/l) = 88.42 (mg/dL)
}
\description{
eGFR
}
\references{
Inker, L. A., Eneanya, N. D., Coresh, J., Tighiouart, H., Wang, D., Sang, Y., ... & Levey, A. S. (2021). New creatinine-and cystatin C–based equations to estimate GFR without race. New England Journal of Medicine, 385(19), 1737-1749.
}
