% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcs_filter.R
\name{filter}
\alias{filter}
\alias{stcs_anonymize}
\alias{stcs_filter_patientkey}
\alias{stcs_remove_emptytab}
\alias{stcs_select_organ}
\alias{stcs_select_table}
\title{Filter/select the STCS dataset}
\usage{
stcs_anonymize(
  stcs,
  vars2rm = c("initials", "dob", "soascaseid", "soasdonorid", "donor_dob")
)

stcs_filter_patientkey(stcs, patientkeys)

stcs_remove_emptytab(stcs)

stcs_select_organ(
  stcs,
  organ = c("Heart", "Islets", "Kidney", "Liver", "Lung", "Pancreas", "Small bowel")
)

stcs_select_table(stcs, tables)
}
\arguments{
\item{stcs}{list. A list of STCS tables}

\item{vars2rm}{chr. A vector of variables to remove.}

\item{patientkeys}{chr. A vector of patientkey to keep.}

\item{organ}{chr. A vector of STCS organ names.}

\item{tables}{chr. A vector of STCS tables name.}
}
\value{
a list of stcs filtered tables.
}
\description{
Filter/select the STCS dataset
}
