% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_lastday.R
\name{impute_lastday}
\alias{impute_lastday}
\title{Impute to the last day of the month/year}
\usage{
impute_lastday(date, dateaccuracy)
}
\arguments{
\item{date}{Date. A vector of date.}

\item{dateaccuracy}{chr. A vector of date accuracy ("Exact date", "Day uncertain", "Day/Month uncertain", "Estimated date")}
}
\value{
a date impute at the last day of the month/year depending on the accuracy.
}
\description{
Impute to the last day of the month/year
}
\examples{
impute_lastday(as.Date("2023-1-12"),"Day uncertain")

}
