% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_intervals.R
\name{merge_intervals}
\alias{merge_intervals}
\title{Merge Interval}
\usage{
merge_intervals(x0, x1, .lag = 0L)
}
\arguments{
\item{x0}{vct. Vector of starting time. Must be complete.}

\item{x1}{vct. Vector of ending time. Must be complete and satisfy \code{x0 <= x1}.}

\item{.lag}{scalar value indicating the time lag when merging interval. See Examples below.}
}
\value{
a \code{data.frame} with 1 row per intervals, without overlapping.
}
\description{
Merge overlapping intervals.
}
\examples{

# return 2 intervals
merge_intervals(c(1, 3, 4), c(2, 4, 5))

# return 1 interval
merge_intervals(c(1, 3, 4), c(2, 4, 5), .lag = 1)
}
