% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimal_key.R
\name{minimal_key}
\alias{minimal_key}
\title{Search the minimal keys in a data frame}
\usage{
minimal_key(x, vmax = 2, key_list = list())
}
\arguments{
\item{x}{data.frame. The data.frame to retrieve the unique key.}

\item{vmax}{int. Default is 2. Define the number of variable that create together a unique key.}

\item{key_list}{list. Use for recursive search.}
}
\value{
a list of variables which are together unique identifiers of the rows.
}
\description{
Recursively identifies minimal keys in a data frame.
}
\examples{

 df <- data.frame(
       id = c(1, 2, 3, 4),
       grp = c("a", "a", "b", "a"),
       val = c(10, 20, 30, 30))

 minimal_key(df)

}
