% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_coding.R
\name{new_coding}
\alias{new_coding}
\title{Recode a variables given the input from a dataframe}
\usage{
new_coding(..., data, .new_var = "new_var")
}
\arguments{
\item{...}{(named) vector.}

\item{data}{a data frame contains the recoding.}

\item{.new_var}{the names of the variable in the recoding data contains the new recoding.}
}
\value{
a vector
}
\description{
Recode a variables given the input from a dataframe
}
\details{
When \code{...} is named the names must match column in \code{data}.
when \code{...} is unnamed, the variables variable in \code{...} and \code{data} are match in order. If \code{data} contains more variables than specified in \code{...} (+1), the functions stop.

If \code{data} contains duplicated values in the subset of column used for the recoding, they are removed and a warning is printed.
}
