% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{noinf_min}
\alias{noinf_min}
\title{Return minimum without infinity}
\usage{
noinf_min(x)
}
\arguments{
\item{x}{a vector different values.}
}
\value{
the minimal value. When all inputs are NA. return NA.
}
\description{
Return minimum without infinity
}
\examples{
noinf_min(c(NA, NA))
noinf_min(c(NA, 4))

}
