% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paste_valuecomment}
\alias{paste_valuecomment}
\title{Merge value and their associated comment}
\usage{
paste_valuecomment(value, comment)
}
\arguments{
\item{value}{a vector different values.}

\item{comment}{a character vector of associated comment}
}
\value{
a character vector of length 1, where values and comments are paste, sorted and separated with \code{" | "}
}
\description{
Merge value and their associated comment
}
\examples{
paste_valuecomment(c("other", "disease A"), c("comment 1", NA))
paste_valuecomment(NA, NA)

}
