% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcs_read.R
\name{read}
\alias{read}
\alias{stcs_read}
\alias{stcs_read1}
\title{Read STCS datastcs}
\usage{
stcs_read(
  dir,
  delim = ",",
  lazy = FALSE,
  progress = FALSE,
  na = "",
  locale = readr::locale(tz = "Europe/Zurich"),
  ...
)

stcs_read1(
  dir,
  filename,
  col_types = NULL,
  delim = ",",
  lazy = FALSE,
  progress = FALSE,
  na = "",
  locale = readr::locale(tz = "Europe/Zurich"),
  ...
)
}
\arguments{
\item{dir}{chr. The directory with ".csv" files.}

\item{delim}{chr. Default is ",". Passed to \code{read_delim}.}

\item{lazy}{lgl. Default is FALSE. Passed to \code{read_delim}.}

\item{progress}{lgl. Default is FALSE. Passed to \code{read_delim}.}

\item{na}{chr. Default is "". Passed to \code{read_delim}.}

\item{locale}{from \code{lubridate::locale()}.}

\item{...}{other arguments passed to \code{read_delim}.}

\item{filename}{chr. The name of the STCS dataset.}

\item{col_types}{Variables type. Passed to \code{read_delim}. See \code{cols}.}
}
\value{
a list of STCS tables.
}
\description{
Read files define in "VariableMetaData.csv"
}
\examples{
\dontrun{
stcs <- stcs_read("path/to/stcs/dataset/")
}

}
\concept{read}
