% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serology_combinaison.R
\name{serology_combinaison}
\alias{serology_combinaison}
\title{Serology combinaison}
\usage{
serology_combinaison(rec_serology, don_serology)
}
\arguments{
\item{rec_serology}{chr. Serology of the recipient. Either "Positive" or Negative.}

\item{don_serology}{chr. Serology of the Donor. Either "Positive" or Negative.}
}
\value{
a character vector.
}
\description{
Recode the serology recipient-donor into: "R+/D+", "R+/D-", "R-/D+", "R-/D-"
}
