% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcs_filter_patientkey.R
\name{stcs_filter_patientkey}
\alias{stcs_filter_patientkey}
\title{Filter the row of all tables given the patientkey}
\usage{
stcs_filter_patientkey(stcs, patientkeys)
}
\arguments{
\item{stcs}{list. A list of STCS tables}

\item{patientkeys}{chr. A vector of \code{patientkey} to keep.}
}
\value{
a list of stcs filtered tables.
}
\description{
Filter the row of all tables given the patientkey
}
\concept{filter}
