% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcs_search.R
\name{stcs_search}
\alias{stcs_search}
\title{Filter for a regex in all character variables.}
\usage{
stcs_search(stcs, pattern, negate = FALSE)
}
\arguments{
\item{stcs}{list. A list of STCS tables}

\item{pattern}{regex. A regex pattern pass to \code{str_detect()}.}

\item{negate}{lgl. Default is \code{FALSE}. Passed to \code{str_detect()}.}
}
\value{
A tibble with the result displayed as a nested table.
}
\description{
Iterate through all dataset, character variables and filter those tables
}
