% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_tables.R
\name{tail_tbl}
\alias{tail_tbl}
\alias{tailored_patientsurvival}
\alias{tailored_organsurvival}
\alias{tailored_transplantationsurvival}
\alias{tailored_psq}
\title{Tailored analysis tables}
\usage{
tailored_patientsurvival(stcs)

tailored_organsurvival(stcs)

tailored_transplantationsurvival(stcs)

tailored_psq(stcs, silent = FALSE)
}
\arguments{
\item{stcs}{A list containing the STCS data frame.}

\item{silent}{lgl. Specify if warning should be returned.}
}
\value{
A data frame.
}
\description{
Tailored analysis tables
}
\details{
\code{tailored_organ()}: PK: \code{organkey}. Baseline and outcome of organs.
\code{tailored_organsurvival()}: PK: \code{patientkey}. Main dates to to compute follow-up time of patients.
\code{tailored_patientsurvival()}: PK: \code{patientkey}. Main dates to to compute follow-up time of patients.
1. \code{patientkey}: Identifier of patients. Source: \code{Patient}.
2. \code{enrollment_date}: Enrollment date. Source: \code{Patient}.
3. \code{deathdate}: Death date. Source: \code{Stop}.
4. \code{first_dropoutdate}: First dropout date. Source: \code{Stop}.
5. \code{first_dropoutdateaccuracy}: Accuracy of the first dropout date. Source: \code{Stop}.
6. \code{last_activedropoutdate}: Last dropout date without \code{backstcs}. Source: \code{Stop}.
7. \code{last_activedropoutdateaccuracy}: Accuracy of the last dropout date without \code{backstcs}. Source: \code{Stop}.
8. \code{last_patlongkeys}: Identifier of the last patient-assessment. Source: \code{PatientLongitudinal}.
9. \code{last_assdate}: Date of the last patient-assessment. Source: \code{PatientLongitudinal}.
10-17. \code{last_[...]_toggle_date}: Assessment date of the last mentioned filled toggle (keep: \code{"Yes"}, \code{"No"}, removed: \code{"Missing"}, \code{NA}). Source: \code{PatientLongitudinal}.
18. \code{extraction_date}: Extraction date. Source: \code{Admin}.
19. \code{last_complete_patientlongitudinal_crf_date}: Assessment date of the latest \code{crf_status} of the patient assessment defined as \code{"Complete"}. Source: \code{PatientLongitudinal}.
20. \code{last_complete_stop_crf_date}: Date (latest between: \code{deathdate}, \code{dropoutdate}, \code{lastalivedate}, \code{backstcsdate}) of the latest \code{crf_status} defined as \code{"Complete"}. Source: \code{Stop}.
21. \code{last_consent_withdrawal_date}: Last consent date when the its consent status is \code{"Withdrawal"}. Source \code{Patient}.
22. \code{confidental data}: Optional.
23. \code{last_complete_psq_crf_date}: Optional. Date of the latest \code{crf_status} of PSQ defined as \code{"Complete"}. Source: \code{PSQ}.

\code{tailored_psq()}: PK: \code{patlongkey}. All PSQ forms in wide format.
}
