% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_psq.R
\name{tailored_psq}
\alias{tailored_psq}
\title{Tailored analysis tables}
\usage{
tailored_psq(stcs, silent = FALSE)
}
\arguments{
\item{stcs}{A list containing the STCS data frame.}

\item{silent}{lgl. Specify if warning should be returned.}
}
\value{
A data frame.
}
\description{
Tailored analysis tables
}
\details{
\code{tailored_psq()}: PK: \code{patlongkey}. All PSQ forms in wide format.
}
