% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_tables.R
\name{tailored_table}
\alias{tailored_table}
\alias{tailored_patientsurvival}
\alias{tailored_organsurvival}
\alias{tailored_transplantationsurvival}
\alias{tailored_psq}
\title{Tailored analysis tables}
\usage{
tailored_patientsurvival(stcs)

tailored_organsurvival(stcs)

tailored_transplantationsurvival(stcs)

tailored_psq(stcs, silent = FALSE)
}
\arguments{
\item{stcs}{A list containing the STCS data frame.}

\item{silent}{lgl. Specify if warning should be returned.}
}
\value{
A data frame.
}
\description{
Tailored analysis tables
}
\details{
\code{tailored_organsurvival()}: PK: \code{organkey}. Main dates to to compute follow-up time of organs.

\itemize{
\item{1.} \code{patientkey}: Identifier of patients. Source: \code{Organ}.
\item{2.} \code{donorkey}: Identifier of donors. Source: \code{Organ}.
\item{3.} \code{soaskey}: Identifier of transplantations. Source: \code{Organ}.
\item{4.} \code{organkey}: Identifier of organs. Source: \code{Organ}.
\item{5.} \code{tpxdate}: Date of transplantation. Source: \code{Organ}.
\item{6.} \code{glo_date}: Graft loss date. Source: \code{Graftloss}.
\item{7.} \code{pnf_date}: Date of PNF (\code{tpxdate} when \code{dgf} is \code{"PNF"}). Source: \code{Organ}.
\item{8.-16.} \code{last_[...]_toggle_date}: Assessment date of the last mentioned filled toggle (keep: \code{"Yes"}, \code{"No"}, removed: \code{"Missing"}, \code{NA}). Source: \code{OrganLongitudinal}.
\item{17.} \code{last_complete_organlongitudinal_crf_date}: Assessment date of the latest \code{crf_status} of the organ assessment defined as \code{"Complete"}. Source: \code{OrganLongitudinal}.
}

\code{tailored_patientsurvival()}: PK: \code{patientkey}. Main dates to to compute follow-up time of patients.

\itemize{

\item{1.} \code{patientkey}: Identifier of patients. Source: \code{Patient}.
\item{2.} \code{enrollment_date}: Enrollment date. Source: \code{Patient}.
\item{3.} \code{deathdate}: Death date. Source: \code{Stop}.
\item{4.} \code{first_dropoutdate}: First dropout date. Source: \code{Stop}.
\item{5.} \code{first_dropoutdateaccuracy}: Accuracy of the first dropout date. Source: \code{Stop}.
\item{6.} \code{last_activedropoutdate}: Last dropout date without \code{backstcs}. Source: \code{Stop}.
\item{7.} \code{last_activedropoutdateaccuracy}: Accuracy of the last dropout date without \code{backstcs}. Source: \code{Stop}.
\item{8.} \code{last_patlongkeys}: Identifier of the last patient-assessment. Source: \code{PatientLongitudinal}.
\item{9.} \code{last_assdate}: Date of the last patient-assessment. Source: \code{PatientLongitudinal}.
\item{10.-17.} \code{last_[...]_toggle_date}: Assessment date of the last mentioned filled toggle (keep: \code{"Yes"}, \code{"No"}, removed: \code{"Missing"}, \code{NA}). Source: \code{PatientLongitudinal}.
\item{18.} \code{extraction_date}: Extraction date. Source: \code{Admin}.
\item{19.} \code{last_complete_patientlongitudinal_crf_date}: Assessment date of the latest \code{crf_status} of the patient assessment defined as \code{"Complete"}. Source: \code{PatientLongitudinal}.
\item{20.} \code{last_complete_stop_crf_date}: Date (latest between: \code{deathdate}, \code{dropoutdate}, \code{lastalivedate}, \code{backstcsdate}) of the latest \code{crf_status} defined as \code{"Complete"}. Source: \code{Stop}.
\item{21.} \code{last_consent_withdrawal_date}: Last consent date when the its consent status is \code{"Withdrawal"}. Source \code{Patient}.
\item{22.} \code{confidental data}: Optional.
\item{23.} \code{last_complete_psq_crf_date}: Optional. Date of the latest \code{crf_status} of PSQ defined as \code{"Complete"}. Source: \code{PSQ}.

}

\code{tailored_psq()}: PK: \code{patlongkey}. All PSQ forms in wide format.
}
\concept{tailored_table}
