% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_tables.R
\name{tailored_tables}
\alias{tailored_tables}
\title{Return a list of tailored tables}
\usage{
tailored_tables(
  stcs,
  tables = c("patientbl", "patientsurvival", "transplantationbl",
    "transplantationsurvival", "organsurvival", "psq")
)
}
\arguments{
\item{stcs}{A list containing the STCS data frame.}

\item{tables}{chr. Define the name of the requested tailored tables. Default is \code{tables = c("patientbl","patientsurvival","transplantationbl","transplantationsurvival","organsurvival","psq"))}.}
}
\value{
a list of dataframe.
}
\description{
Return a list of tailored tables
}
