% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_transplantationsurvival.R
\name{tailored_transplantationsurvival}
\alias{tailored_transplantationsurvival}
\title{Tailored analysis tables with main dates related to transplantation}
\usage{
tailored_transplantationsurvival(stcs)
}
\arguments{
\item{stcs}{A list containing the STCS data frame.}
}
\value{
A data frame.
}
\description{
Tailored analysis tables with main dates related to transplantation
}
\details{
\code{tailored_transplantationsurvival()}: PK: \code{soaskey}.

\itemize{

\item{1.} \code{soaskey}: Identifier of transplantations. Source: \code{Transplantation}.
\item{2.} \code{patientkey}: Identifier of patients. Source: \code{Transplantation}.
\item{3.} \code{donorkey}: Identifier of patients. Source: \code{Transplantation}.
\item{4.} \code{listing_date}: Date of listing. Source: \code{Transplantation}.
\item{5.} \code{removal_date}: Date of removal. Source: \code{Transplantation}.
\item{6.} \code{soas_tpxdate}: (Latest) Date of transplantation. Source: \code{Transplantation}.
\item{7.} \code{next_tpx_tpxdate}: Date of the next transplantation (all \code{tpx}). Source: \code{Transplantation}.
\item{8.} \code{next_tpxspecific_tpxdate}: Date of the next transplantation (\code{tpx} specific). Source: \code{Transplantation}.
\item{9.} \code{last_complete_biosample_crf_date}: (Optional) Date of latest completed biosample \code{crf_status}. Source: \code{Biosample}.

}
}
