% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truemissing.R
\name{truemissing}
\alias{truemissing}
\alias{is_truemissing}
\alias{is_truemissing.character}
\alias{is_truemissing.Date}
\alias{is_truemissing.factor}
\alias{is_truemissing.integer}
\alias{is_truemissing.logical}
\alias{is_truemissing.numeric}
\alias{truemissing_to_na}
\alias{is_truemissing.POSIXct}
\title{Confirmed missing values}
\usage{
is_truemissing(x)

\method{is_truemissing}{character}(x)

\method{is_truemissing}{Date}(x)

\method{is_truemissing}{factor}(x)

\method{is_truemissing}{integer}(x)

\method{is_truemissing}{logical}(x)

\method{is_truemissing}{numeric}(x)

truemissing_to_na(x)

\method{is_truemissing}{POSIXct}(x)
}
\arguments{
\item{x}{a vector. Logical are not implemented.}
}
\value{
a vector of logical
}
\description{
Handles confirmed missing values
}
\details{
\code{is_truemissing()}: return a logical
\code{truemissing_to_na()}: replace confirmed missing by NA's
}
\concept{truemissing}
