% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_metadata.R
\name{validate_metadata}
\alias{validate_metadata}
\title{Validate the VariableMetadata}
\usage{
validate_metadata(stcs)
}
\arguments{
\item{stcs}{a list containing the STCS dataset.}
}
\value{
a warining with variable mismatches or the \code{stcs} object.
}
\description{
Validate the VariableMetadata
}
