% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_available.R
\name{var_available}
\alias{var_available}
\title{Print variables which are available given your current dataset}
\usage{
var_available(data, stcs)
}
\arguments{
\item{data}{dataframe. A data frame with one or more STCS keys}

\item{stcs}{list. A list of STCS tables}
}
\value{
Invisible input data frame.

inputed data
}
\description{
Print variables which are available given your current dataset
}
