% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec_names}
\alias{vec_names}
\title{Assigned names to a vector}
\usage{
vec_names(x, nm = NULL)
}
\arguments{
\item{x}{a vector.}

\item{nm}{a character vector. Default is \code{NULL}.}
}
\value{
a named vector.
}
\description{
Assigned names to a vector
}
\details{
If \code{nm = NULL}, the names are the value of the vector (for the unnamed element.). \code{vec_names} is different from \code{rlang::set_names}: each each element of \code{x}, it assigns the values of \code{x} as its name when its \code{names(x)} is missing (values is \code{""}).
}
