% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which.pmin.R
\name{whichp}
\alias{whichp}
\alias{which.pmin}
\alias{which.pmin_chr}
\title{Indicate the index of the row-wise minimal value}
\usage{
which.pmin(..., silent = F)

which.pmin_chr(..., silent = F)
}
\arguments{
\item{...}{Vectors of the same type where which.min is applied.}

\item{silent}{a logical indicating if the function should return a warning in case of duplicated minimal.}
}
\value{
an integer vector indicating the row-wise index of the (first) minimal value
}
\description{
Indicate the index of the row-wise minimal value
}
\details{
See original idea https://github.com/radiant-rstats/radiant.data/blob/master/R/radiant.R
}
\concept{whichp}
