% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stcs_write.R
\name{write}
\alias{write}
\alias{stcs_write_csv}
\alias{stcs_write_xlsx}
\title{Write a list of STCS dataset into CSV file}
\usage{
stcs_write_csv(stcs, dir, delim = ", ", progress = FALSE, na = "", ...)

stcs_write_xlsx(stcs, dir)
}
\arguments{
\item{stcs}{a list of STCS datasets}

\item{dir}{chr. Path of the directory containing the exported files data. Dir is created if it does not exists.}

\item{delim}{the column delimiting character passed to \code{write_delim()}. Default is ", ".}

\item{progress}{Function or lgl. Argument pass to \code{write_delim()} Default is FALSE.}

\item{na}{character string used for missing values. Passed to \code{write_delim()}. Default is "".}

\item{...}{other arguments passed to \code{write_delim()}.}
}
\description{
Write a list of STCS dataset into CSV file
}
\examples{
\dontrun{
stcs_write(dir = "path/to/empty/folder", stcs = stcs)
}

}
\concept{write}
