% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_infsite.R
\name{categorize_infsite}
\alias{categorize_infsite}
\title{Extract infection site data}
\usage{
categorize_infsite(data, stcs, .diseasekey = "diseasekey")
}
\arguments{
\item{data}{data.frame. A data frame containing the variables describe in \code{.patientkey} and \code{.date}}

\item{stcs}{list. The stcs object.}

\item{.diseasekey}{chr. The column name containing \code{diseasekey}.}
}
\value{
a data frame containing \code{"infsite"},  \code{"comment"}, \code{"n_occurence"}.
}
\description{
Extract infection site data
}
