% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorize_pathogenspecies.R
\name{categorize_pathogenspecies}
\alias{categorize_pathogenspecies}
\title{Extract infection pathogen species data for recoding}
\usage{
categorize_pathogenspecies(data, stcs, .diseasekey = "diseasekey")
}
\arguments{
\item{data}{data.frame. A data frame containing the variables describe in \code{.patientkey} and \code{.date}}

\item{stcs}{list. The stcs object.}

\item{.diseasekey}{chr. The column name containing diseasekey}
}
\value{
a data frame containing \code{"patdiagnosis"}, \code{"pathogen_species"}, \code{"comment"}, \code{"n_occurence"}.
}
\description{
Extract infection pathogen species data for recoding
}
