% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egfr.R
\name{pediatric_egfr_2009}
\alias{pediatric_egfr_2009}
\title{Swchartz bedside pediatric eGFR formula}
\usage{
pediatric_egfr_2009(crea_mumoll, height)
}
\arguments{
\item{crea_mumoll}{Numeric. Creatinin in \eqn{\mu}mol/l}

\item{height}{Numeric. Height in meters.}
}
\value{
eGFR in mL/min/1.73m\eqn{^2} as described in https://pubmed.ncbi.nlm.nih.gov/19158356/. It corresponds to the bedside formula: 0.413*height_cm/crea_mgdl. The conversion factor of creatinin is 1 (\eqn{\mu}mol/l) = 88.42 (mg/dL).
}
\description{
Swchartz bedside pediatric eGFR formula
}
\references{
Schwartz, G. J., Schneider, M. F., Maier, P. S., Moxey-Mims, M., Dharnidharka, V. R., Warady, B. A., ... & Muñoz, A. (2012). Improved equations estimating GFR in children with chronic kidney disease using an immunonephelometric determination of cystatin C. Kidney international, 82(4), 445-453.
}
