% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_tables.R
\name{tail_tbl}
\alias{tail_tbl}
\alias{tailored_organ}
\alias{tailored_patientsurvival}
\alias{tailored_organsurvival}
\alias{tailored_transplantationsurvival}
\alias{tailored_psq}
\title{Tailored analysis tables}
\usage{
tailored_organ(stcs)

tailored_patientsurvival(stcs)

tailored_organsurvival(stcs)

tailored_transplantationsurvival(stcs)

tailored_psq(stcs, silent = FALSE)
}
\arguments{
\item{stcs}{A list containing the STCS data frame.}

\item{silent}{lgl. Specify if warning should be returned.}
}
\value{
A data frame.
}
\description{
Tailored analysis tables
}
\details{
\code{tailored_organ()}: PK: \code{organkey}. Baseline and outcome of organs.
\code{tailored_patientsurvival()}: PK: \code{patientkey}. Main date for survival of the patients.
\code{tailored_psq()}: PK: \code{patlongkey}. All PSQ forms in wide format.
}
