% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_tables.R
\name{tail_tbl}
\alias{tail_tbl}
\alias{tailored_organ}
\alias{tailored_patientsurvival}
\alias{tailored_psq}
\title{Tailored analysis tables}
\usage{
tailored_organ(stcs)

tailored_patientsurvival(stcs)

tailored_psq(stcs)
}
\arguments{
\item{stcs}{A list containing the STCS data frame.}
}
\value{
A data frame.
}
\description{
Tailored analysis tables
}
\details{
\code{tailored_organ()}: PK: \code{organkey}. Baseline and outcome of organs.
}
