% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailored_organsurvival.R,
%   R/tailored_patientbl.R, R/tailored_patientsurvival.R, R/tailored_psq.R,
%   R/tailored_transplantationbl.R, R/tailored_transplantationsurvival.R
\name{tailored_table}
\alias{tailored_table}
\alias{tailored_organsurvival}
\alias{tailored_patientbl}
\alias{tailored_patientsurvival}
\alias{tailored_psq}
\alias{tailored_transplantationbl}
\alias{tailored_transplantationsurvival}
\title{Tailored analysis tables with main dates related to organs.}
\usage{
tailored_organsurvival(stcs)

tailored_patientbl(stcs)

tailored_patientsurvival(stcs)

tailored_psq(stcs, silent = FALSE)

tailored_transplantationbl(stcs)

tailored_transplantationsurvival(stcs)
}
\arguments{
\item{stcs}{A list containing the STCS data frame.}

\item{silent}{lgl. Specify if warning should be returned.}
}
\value{
A data frame.

A data frame.

A data frame.

A data frame.

A data frame.

A data frame.
}
\description{
Tailored analysis tables with main dates related to organs.

Tailored analysis tables with baseline patient data.

Tailored analysis tables with main dates related to patients

Tailored analysis tables

Tailored analysis tables with with baseline transplantation data.

Tailored analysis tables with main dates related to transplantation
}
\details{
\code{tailored_organsurvival()}: PK: \code{organkey}.

\itemize{
\item{1.} \code{patientkey}: Identifier of patients. Source: \code{Organ}.
\item{2.} \code{donorkey}: Identifier of donors. Source: \code{Organ}.
\item{3.} \code{soaskey}: Identifier of transplantations. Source: \code{Organ}.
\item{4.} \code{organkey}: Identifier of organs. Source: \code{Organ}.
\item{5.} \code{tpxdate}: Date of transplantation. Source: \code{Organ}.
\item{6.} \code{glo_date}: Graft loss date. Source: \code{Graftloss}.
\item{7.} \code{pnf_date}: Date of PNF (\code{tpxdate} when \code{dgf} is \code{"PNF"}). Source: \code{Organ}.
\item{8.-16.} \code{last_[...]_toggle_date}: Assessment date of the last mentioned filled toggle (keep: \code{"Yes"}, \code{"No"}, removed: \code{"Missing"}, \code{NA}). Source: \code{OrganLongitudinal}.
\item{17.} \code{last_complete_organlongitudinal_crf_date}: Assessment date of the latest \code{crf_status} of the organ assessment defined as \code{"Complete"}. Source: \code{OrganLongitudinal}.
}

\code{tailored_patientbl()}: PK: \code{patientkey}.

\itemize{
\item{1.} \code{patientkey}: Identifier of patients. Source: \code{Patient}.
\item{2.} \code{patid}: STCS Identifier of patients. Source: \code{Patient}.
\item{3.} \code{enrollment_age}: Patient age at enrollment. Source: \code{Patient}.
\item{4.} \code{sex}: Patient sex. Source: \code{Patient}.
\item{5.} \code{bg}: Patient blood group. Source: \code{Patient}.
\item{6.} \code{pre_tpx_count}: Number of pre-STSC transplantation. Source: \code{Patient}.
\item{7.} \code{pre_organ_count}: Number of pre-STSC organ. Source: \code{Patient}.
\item{8.} \code{weight}: Weight of patient at enrollment. Source: \code{PatientLongitudinal}.
\item{9.} \code{height}: Height of patient at enrollment. Source: \code{PatientLongitudinal}.
\item{10.-24.} \code{hla...}: (Optional) Patient HLA. Source: \code{PatientHLA}.
}

\code{tailored_patientsurvival()}: PK: \code{patientkey}.

\itemize{

\item{1.} \code{patientkey}: Identifier of patients. Source: \code{Patient}.
\item{2.} \code{enrollment_date}: Enrollment date. Source: \code{Patient}.
\item{3.} \code{deathdate}: Death date. Source: \code{Stop}.
\item{4.} \code{first_dropoutdate}: First dropout date. Source: \code{Stop}.
\item{5.} \code{first_dropoutdateaccuracy}: Accuracy of the first dropout date. Source: \code{Stop}.
\item{6.} \code{last_activedropoutdate}: Last dropout date without \code{backstcs}. Source: \code{Stop}.
\item{7.} \code{last_activedropoutdateaccuracy}: Accuracy of the last dropout date without \code{backstcs}. Source: \code{Stop}.
\item{8.} \code{last_patlongkeys}: Identifier of the last patient-assessment. Source: \code{PatientLongitudinal}.
\item{9.} \code{last_assdate}: Date of the last patient-assessment. Source: \code{PatientLongitudinal}.
\item{10.-17.} \code{last_[...]_toggle_date}: Assessment date of the last mentioned filled toggle (keep: \code{"Yes"}, \code{"No"}, removed: \code{"Missing"}, \code{NA}). Source: \code{PatientLongitudinal}.
\item{18.} \code{extraction_date}: Extraction date. Source: \code{Admin}.
\item{19.} \code{last_complete_patientlongitudinal_crf_date}: Assessment date of the latest \code{crf_status} of the patient assessment defined as \code{"Complete"}. Source: \code{PatientLongitudinal}.
\item{20.} \code{last_complete_stop_crf_date}: Date (latest between: \code{deathdate}, \code{dropoutdate}, \code{lastalivedate}, \code{backstcsdate}) of the latest \code{crf_status} defined as \code{"Complete"}. Source: \code{Stop}.
\item{21.} \code{last_consent_withdrawal_date}: Last consent date when the its consent status is \code{"Withdrawal"}. Source \code{Patient}.
\item{22.} \code{confidental data}: Optional.
\item{23.} \code{last_complete_psq_crf_date}: Optional. Date of the latest \code{crf_status} of PSQ defined as \code{"Complete"}. Source: \code{PSQ}.

}

\code{tailored_psq()}: PK: \code{patlongkey}. All PSQ forms in wide format.

\code{tailored_transplantationbl()}: PK: \code{soaskey}.


\itemize{

\item{1.} \code{soaskey}: Identifier of transplantations. Source: \code{Transplantation}.
\item{2.} \code{patientkey}: Identifier of patients. Source: \code{Transplantation}.
\item{3.} \code{donorkey}: Identifier of donors. Source: \code{Transplantation}.
\item{4.} \code{organ_count}: Number of organs in the tpx. Source: \code{Transplantation}.
\item{5.} \code{tpx}: List of organs. Source: \code{Transplantation}.
\item{6.} \code{tpx_resec}: transplantation level Re, Sec. Source: \code{Transplantation}.
\item{7.} \code{pre_tpxspecific_count}: Number of pre-transplantation (\code{tpx} specific). Source: \code{Transplantation}.
\item{8.} \code{tpxspecific_order}: Order of transplantion (\code{tpx} specific). Source : \code{Transplantation}.
\item{9.} \code{tpx_order}: Order of transplantion. Source: \code{Transplantation}.
\item{10.} \code{tpx_stcsorder}: Order of transplantion within the STCS. Source: \code{Transplantation}.
\item{11.} \code{abocomp}: Blood groups compatibility. Source: \code{Transplantation}.
\item{12.-18.} \code{hla...mismatch}: HLA mismatch. Source: \code{Transplantation}.
\item{19.} \code{sumhlamismatch}: Sum of mismatch. Source: \code{Transplantation}.
\item{20.} \code{hlaavailable}: Number of HLA mismatch available. Source: \code{Transplantation}.
\item{21.} \code{procedure_provider}: Procedure provider. Source: \code{Transplantation}.
\item{22.} \code{dontype}: Donor type. Source: \code{Donor}.
\item{23.} \code{donrelsubtype}: Donor subtype. Source: \code{Donor}.
\item{24.} \code{identical_twins}: Identical twins. Source: \code{Donor}.
\item{25.} \code{kpd}: Kidney paired donation. Source: \code{Donor}.
\item{26.} \code{altruistic}: Altruistic donation. Source: \code{Donor}.
\item{27.} \code{don_bg}: Blood-group of the donor. Source: \code{Donor}.
\item{28.} \code{don_sex}: Sex of the donor. Source: \code{Donor}.
\item{29.} \code{doncod}: Cause of donor death. Source: \code{Donor}.
\item{30.} \code{doncod_comment}: Comment on cause of donor death. Source: \code{Donor}.
\item{31.} \code{donage}: Age of donor. Source: \code{Donor}.
\item{32.} \code{donage_complete}: Calculated age of the donor. Source: \code{Donor}.
\item{33.-43.} \code{rec_anti...}: (Optional) Recipient Serology. Source: \code{PatientSerology}.
\item{44.-54.} \code{don_anti...}: (Optional) Donor Serology. Source: \code{DonorSerology}.
\item{55.-65.} \code{combi_anti...}: (Optional) Serology combination. Source: derived.

}

\code{tailored_transplantationsurvival()}: PK: \code{soaskey}.

\itemize{

\item{1.} \code{soaskey}: Identifier of transplantations. Source: \code{Transplantation}.
\item{2.} \code{patientkey}: Identifier of patients. Source: \code{Transplantation}.
\item{3.} \code{donorkey}: Identifier of patients. Source: \code{Transplantation}.
\item{4.} \code{listing_date}: Date of listing. Source: \code{Transplantation}.
\item{5.} \code{removal_date}: Date of removal. Source: \code{Transplantation}.
\item{6.} \code{soas_tpxdate}: (Latest) Date of transplantation. Source: \code{Transplantation}.
\item{7.} \code{next_tpx_tpxdate}: Date of the next transplantation (all \code{tpx}). Source: \code{Transplantation}.
\item{8.} \code{next_tpxspecific_tpxdate}: Date of the next transplantation (\code{tpx} specific). Source: \code{Transplantation}.
\item{9.} \code{last_complete_biosample_crf_date}: (Optional) Date of latest completed biosample \code{crf_status}. Source: \code{Biosample}.

}
}
